+++ src/webapp/src/views/Orchestration.vue
@@ -1,5 +1,5 @@
 <script>
-import { mapActions, mapGetters } from 'vuex'
+import { mapGetters } from 'vuex'
 
 import Airflow from '@/components/orchestration/Airflow'
 import RouterViewLayout from '@/views/RouterViewLayout'
@@ -10,22 +10,10 @@ export default {
     Airflow,
     RouterViewLayout
   },
-  created() {
-    this.$store.dispatch('plugins/getInstalledPlugins')
-  },
   computed: {
     ...mapGetters('plugins', ['getIsPluginInstalled', 'getIsInstallingPlugin']),
     getIsInstallingAirflow() {
       return this.getIsInstallingPlugin('orchestrators', 'airflow')
-    }
-  },
-  methods: {
-    ...mapActions('plugins', ['addPlugin', 'installPlugin']),
-    installAirflow() {
-      const payload = { pluginType: 'orchestrators', name: 'airflow' }
-      this.addPlugin(payload).then(() => {
-        this.installPlugin(payload)
-      })
     }
   }
 }
@@ -49,26 +37,26 @@ export default {
         <div class="columns">
           <div class="column">
             <div class="content">
+              <template v-if="getIsInstallingAirflow">
+                <p>
+                  Airflow is Meltano's current orchestrator. Learn what's
+                  possible in the
+                  <a
+                    target="_blank"
+                    href="https://www.meltano.com/docs/meltano-cli.html#orchestration"
+                    >Meltano Airflow</a
+                  >
+                  documentation.
+                </p>
+                <hr />
+                <p class="is-italic has-text-centered">
+                  Airflow installation can take a few minutes.
+                </p>
+              </template>
+
               <p>
-                Airflow is Meltano's current orchestrator. Learn what's possible
-                in the
-                <a
-                  target="_blank"
-                  href="https://www.meltano.com/docs/meltano-cli.html#orchestration"
-                  >Meltano Airflow</a
-                >
-                documentation.
+                <progress class="progress is-small is-info"></progress>
               </p>
-              <p v-if="getIsInstallingAirflow" class="is-italic">
-                Airflow installation can take a few minutes.
-              </p>
-              <a
-                class="button is-interactive-primary"
-                :class="{ 'is-loading': getIsInstallingAirflow }"
-                @click="installAirflow"
-              >
-                Install Airflow
-              </a>
             </div>
           </div>
         </div>